Removed Location Desialization Fix  caused by version 1.0 a year ago
I figured it was time to remove the Location Desialization Fix that was
caused close to a year ago. I am just going to assume that anyone that
has version 1.0 and finally decides to update the plugin can do with the
loss of data since 1.0 was over a year ago.